backend emails refactor for tenants
